requirements.txt
h5py==3.9.0
matplotlib==3.8.2
numpy==1.26.4
opencv_python==4.8.1.78
Pillow==10.3.0
plyfile==1.0.3
pytorch3d==0.7.5
pytorch_fid==0.3.0
scikit_learn==1.3.2
scipy==1.13.0
setuptools==69.0.3
torch==2.1.1
torchvision==0.16.1
tqdm==4.65.0
warp==1.0.4
warp_lang==0.10.1
